requirements.txt
tensorflow>=2.10.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.20.0
huggingface-hub>=0.16.0
einops>=0.6.0
pytest>=7.3.0
pytest-cov>=4.1.0
